Cut ssl from localhost
A CA localhost certificate was used by using a DNS entry pointing to localhost. This is considered bad-practice so we drop this certificate.
As an alternative we could have chosen to generate a localhost on install but this wouldn't be trusted by all browsers (which is actually a good thing). So we choose to use a plain channel because browsers should consider the localhost as a securecontext.